love/hate now works`

+++ site/pic.php
@@ -14,14 +14,41 @@ if (!isset($_GET['id']))
 $container = $CONFIG->Container();
 try {
 	$obj = $container->DataObject($_GET['id']);
+	$delete_at = $obj->extra_headers['X-Delete-At'];
+	// how many seconds till expired?
+	$diff = $delete_at - time();
 
 	/**
 	 * handle Like, Dislike
 	 */
+	if (isset($_POST['Love'])) {
+		// love it! add time
+		$delete_at += $CONFIG->love_extra;
+		$obj->updateMetadata(array(
+			'X-Delete-At' => $delete_at
+		));
+		$LOVED=TRUE;
+	}
 	if (isset($_POST['Hate'])) {
-		$obj->delete();
-		header('Location: http://'.$_SERVER['HTTP_HOST']);
-		exit;
+		/*
+		error_log(sprintf('Hate[%s]: %d seconds left',
+			$_GET['id'], $diff));
+		*/
+
+		// if less than a minute, delete it
+		if ($diff < 60) {
+			$obj->delete();
+			header('Location: http://'.$_SERVER['HTTP_HOST']);
+			exit;
+		}
+		else {
+			// otherwise, cut the remaining time in half
+			$delete_at = time() + round($diff/2);
+			$obj->updateMetadata(array(
+				'X-Delete-At' => $delete_at
+			));
+			$HATED=TRUE;
+		}
 	}
 
 	/**
@@ -30,7 +57,9 @@ try {
 	$PIC = new \stdClass;
 	$PIC->name = $obj->Name();
 	$PIC->url = $obj->PublicURL();
+	$PIC->expiration = $delete_at;
 } catch (Exception $e) {
+print_r($e);
 	header('HTTP/1.1 404 NOT FOUND');
 	$TITLE = 'Pictr - NOT FOUND OMG';
 	$PIC = FALSE;
